refactor node for plan
